Refactor products code (#16)
- Created helper function renderProducts to address code duplication
- Moved ProductsDetailedScreen.js and ProductsScreen onto a products folder, similar to stores
- Changed Products component
