+++ components/StoreProducts.js
@@ -1,10 +1,10 @@
+import StoreCard from './StoreCard';
 import React from 'react';
 import { View } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
+import Product from './Product';
 
 import { Button, Title } from '../styles/shared';
-import Product from './Product';
-import StoreCard from './StoreCard';
 
 function filterFruit(product) {
   if (product) {
@@ -27,8 +27,41 @@ class StoreProducts extends React.Component {
     });
   };
 
+  renderProducts = (filterType, productType) => {
+    const { navigation, store, products } = this.props;
+    return (
+      <View>
+        <View flexDirection="row">
+          <Title>{productType}</Title>
+          <Button
+            onPress={() =>
+              navigation.navigate('Products', {
+                products: products.filter(filterType),
+                navigation,
+                productType,
+                store
+              })
+            }>
+            <Title>See all</Title>
+          </Button>
+        </View>
+        <ScrollView horizontal showsHorizontalScrollIndicator={false}>
+          {products.filter(filterType).map(product => (
+            // TODO See if there is a better way to pass the props over to a component
+            <Product
+              key={product.id}
+              product={product}
+              navigation={navigation}
+              store={store}
+            />
+          ))}
+        </ScrollView>
+      </View>
+    );
+  };
+
   render() {
-    const { navigation, products, store } = this.props;
+    const { store } = this.props;
     return (
       <View>
         <StoreCard
@@ -36,65 +69,9 @@ class StoreProducts extends React.Component {
           key={store.id}
           callBack={() => this.detailedStoreTransition(store)}
         />
-        {/* TODO @tommypoa can probably make this (e.g Fruits title + fruits products) a sub-component and re-use it in AllProducts */}
         {/* Display fruits available at this store */}
-        <View flexDirection="row">
-          <Title>Fruits</Title>
-          <Button
-            onPress={() =>
-              navigation.navigate('Products', {
-                products: products.filter(filterFruit),
-                navigation: this.props.navigation,
-                productType: 'Fruits',
-                store
-              })
-            }>
-            <Title>See all</Title>
-          </Button>
-        </View>
-        <ScrollView horizontal showsHorizontalScrollIndicator={false}>
-          {products.filter(filterFruit).map(product => (
-            <Button
-              key={product.id}
-              onPress={() =>
-                navigation.navigate('ProductsDetailed', {
-                  currentProduct: product,
-                  store
-                })
-              }>
-              <Product product={product} />
-            </Button>
-          ))}
-        </ScrollView>
-        {/* Display vegetables available at this store */}
-        <View flexDirection="row">
-          <Title>Veggies</Title>
-          <Button
-            onPress={() =>
-              navigation.navigate('Products', {
-                products: products.filter(filterVegetables),
-                navigation,
-                productType: 'Vegetables',
-                store
-              })
-            }>
-            <Title>See all</Title>
-          </Button>
-        </View>
-        <ScrollView horizontal showsHorizontalScrollIndicator={false}>
-          {products.filter(filterVegetables).map(product => (
-            <Button
-              key={product.id}
-              onPress={() =>
-                navigation.navigate('ProductsDetailed', {
-                  currentProduct: product,
-                  store
-                })
-              }>
-              <Product product={product} />
-            </Button>
-          ))}
-        </ScrollView>
+        <View>{this.renderProducts(filterFruit, 'Fruit')}</View>
+        <View>{this.renderProducts(filterVegetables, 'Vegetables')}</View>
       </View>
     );
   }
