+++ screens/products/ProductsDetailedScreen.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Image, View } from 'react-native';
 
-import ProductInfo from '../components/ProductInfo';
-import { ImageContainer } from '../styles/products';
-import { Title } from '../styles/shared';
+import ProductInfo from '../../components/ProductInfo';
+import { ImageContainer } from '../../styles/products';
+import { Title } from '../../styles/shared';
 
 class ProductsDetailedScreen extends React.Component {
   constructor(props) {
@@ -20,7 +20,7 @@ class ProductsDetailedScreen extends React.Component {
           <ProductInfo product={currentProduct} />
           <View style={{ marginTop: 40 }}>
             <Image
-              source={require('../assets/images/robot-dev.png')}
+              source={require('../../assets/images/robot-dev.png')}
               style={{ width: 80, height: 80, borderRadius: 80 / 2 }}
             />
           </View>
